dz-modules/packages/dz-hypergraph/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dz-hypergraph"
version = "0.1.0"
description = "Discovery Zero hypergraph layer with production tools"
requires-python = ">=3.12"
dependencies = [
  "gaia-lang",
  "pydantic>=2.0",
  "typer[all]>=0.12",
  "numpy>=1.26",
  "httpx>=0.27",
  "pydantic-settings>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dz_hypergraph"]

[tool.hatch.build.targets.wheel.force-include]
"src/dz_hypergraph/skills" = "dz_hypergraph/skills"
